creates new more page with new styling (#382)
